add worker to bitswap for reproviding new blocks
